refactor: extract config tests to sibling file
